Add ControllerAdvice to a new class, fixed some bugs and enhanced some code all over the project.

+++ CodeforcesSmallMimic/src/main/java/com/zerowhisper/codeforcessmallmimic/dto/ProblemDto.java
@@ -1,15 +1,15 @@
 package com.zerowhisper.codeforcessmallmimic.dto;
 
+import com.zerowhisper.codeforcessmallmimic.entity.UserAccount;
 import lombok.Data;
 
 @Data
 public class ProblemDto {
     // DONE
-
     private String title;
     private String description;
     private String testCases;
     private String expectedOutput;
-    private Integer memoryLimit;
-    private Integer timeLimit;
+    private Double memoryLimit;
+    private Double timeLimit;
 }
